Reworked gpio access to work with current kernel versions.
Newer kernels error if you export a GPIO that's already exported, and we used to rely on being able to do this. Now gpios are centralized in the main module, and all the other modules just share that hash.
This is temporary, I want to break that out from the main module as I don't like raspi-specific stuff in there, but this makes it work with the new kernels instead of crashing at least.