Changed ext_manager to the more generic module_manager.
My plan is to make everything more dynamically loadable modules, so you can load multiple different UI output modules for example, to get text on an LCD and also read out sort of like a built-in screen reader.
First step was to make the module loading system used for extension modules more generic and add a 'category' system to it, which is now done. Still only used for external modules right now, but the framework is in place to move other types of modules into it.