Added screen reader functionality in screen_reader.
This can be used if you want to build a timer without a screen at all, or for use without needing to/being able to see the screen. All content that isn't otherwise read out is read out this way, in a way that makes the UI as easy to use as possible.